% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R, R/mult.R
\name{mult}
\alias{mult}
\title{Multiply two numbers}
\usage{
mult(x, y)

mult(4, 5)
mult(5, 3)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
The product of \code{x} and \code{y}.

product The product of multiplying x and y
}
\description{
Multiply two numbers

Multiply together two numbers
}
\examples{
mult(2, 5)

}
